Add simple output option for time conversion
This commit introduces a new "simple" option which allows users to print
just the converted time when using the "time" command. Additionally, a
new function 'convert_time_string' is introduced to convert the time
string for 'simple' mode. Code is also included to handle and return an
error message if an invalid time string is provided.